requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
pillow==10.4.0
pytesseract==0.3.13
